Move remaining deframer buffer types to buffers.rs
